+++ docs/script/search_index.js
@@ -2616,6 +2616,12 @@ window.esdocSearchIndex = [
     "member"
   ],
   [
+    "src/pdfbuilder.js~pdfbuilder#highlightinvalidcomponents",
+    "class/src/PDFBuilder.js~PDFBuilder.html#instance-method-highlightInvalidComponents",
+    "src/PDFBuilder.js~PDFBuilder#highlightInvalidComponents",
+    "method"
+  ],
+  [
     "src/pdfbuilder.js~pdfbuilder#init",
     "class/src/PDFBuilder.js~PDFBuilder.html#instance-method-init",
     "src/PDFBuilder.js~PDFBuilder#init",
@@ -2668,6 +2674,12 @@ window.esdocSearchIndex = [
     "class/src/PDFBuilder.js~PDFBuilder.html#instance-method-render",
     "src/PDFBuilder.js~PDFBuilder#render",
     "method"
+  ],
+  [
+    "src/pdfbuilder.js~pdfbuilder#repeatablepaths",
+    "class/src/PDFBuilder.js~PDFBuilder.html#instance-member-repeatablePaths",
+    "src/PDFBuilder.js~PDFBuilder#repeatablePaths",
+    "member"
   ],
   [
     "src/pdfbuilder.js~pdfbuilder#setuploaderror",
@@ -3390,6 +3402,12 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/webformbuilder.js~webformbuilder#findrepeatablepaths",
+    "class/src/WebformBuilder.js~WebformBuilder.html#instance-method-findRepeatablePaths",
+    "src/WebformBuilder.js~WebformBuilder#findRepeatablePaths",
+    "method"
+  ],
+  [
     "src/webformbuilder.js~webformbuilder#form",
     "class/src/WebformBuilder.js~WebformBuilder.html#instance-get-form",
     "src/WebformBuilder.js~WebformBuilder#form",
@@ -3502,6 +3520,12 @@ window.esdocSearchIndex = [
     "class/src/WebformBuilder.js~WebformBuilder.html#instance-method-render",
     "src/WebformBuilder.js~WebformBuilder#render",
     "method"
+  ],
+  [
+    "src/webformbuilder.js~webformbuilder#repeatablepaths",
+    "class/src/WebformBuilder.js~WebformBuilder.html#instance-member-repeatablePaths",
+    "src/WebformBuilder.js~WebformBuilder#repeatablePaths",
+    "member"
   ],
   [
     "src/webformbuilder.js~webformbuilder#savecomponent",
@@ -4149,6 +4173,12 @@ window.esdocSearchIndex = [
     "src/components/_classes/component/component.js~component#_parentdisabled",
     "class/src/components/_classes/component/Component.js~Component.html#instance-member-_parentDisabled",
     "src/components/_classes/component/Component.js~Component#_parentDisabled",
+    "member"
+  ],
+  [
+    "src/components/_classes/component/component.js~component#_parentpath",
+    "class/src/components/_classes/component/Component.js~Component.html#instance-member-_parentPath",
+    "src/components/_classes/component/Component.js~Component#_parentPath",
     "member"
   ],
   [
@@ -5703,6 +5733,12 @@ window.esdocSearchIndex = [
     "src/components/_classes/input/input.js~input#getwidget",
     "class/src/components/_classes/input/Input.js~Input.html#instance-method-getWidget",
     "src/components/_classes/input/Input.js~Input#getWidget",
+    "method"
+  ],
+  [
+    "src/components/_classes/input/input.js~input#getwordcount",
+    "class/src/components/_classes/input/Input.js~Input.html#instance-method-getWordCount",
+    "src/components/_classes/input/Input.js~Input#getWordCount",
     "method"
   ],
   [
@@ -8676,6 +8712,12 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/components/editgrid/editgrid.js~editgridcomponent#baseremoverow",
+    "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-method-baseRemoveRow",
+    "src/components/editgrid/EditGrid.js~EditGridComponent#baseRemoveRow",
+    "method"
+  ],
+  [
     "src/components/editgrid/editgrid.js~editgridcomponent#cancelrow",
     "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-method-cancelRow",
     "src/components/editgrid/EditGrid.js~EditGridComponent#cancelRow",
@@ -8709,12 +8751,6 @@ window.esdocSearchIndex = [
     "src/components/editgrid/editgrid.js~editgridcomponent#clearerrors",
     "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-method-clearErrors",
     "src/components/editgrid/EditGrid.js~EditGridComponent#clearErrors",
-    "method"
-  ],
-  [
-    "src/components/editgrid/editgrid.js~editgridcomponent#clearonhide",
-    "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-method-clearOnHide",
-    "src/components/editgrid/EditGrid.js~EditGridComponent#clearOnHide",
     "method"
   ],
   [
@@ -8842,12 +8878,6 @@ window.esdocSearchIndex = [
     "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-get-minLength",
     "src/components/editgrid/EditGrid.js~EditGridComponent#minLength",
     "member"
-  ],
-  [
-    "src/components/editgrid/editgrid.js~editgridcomponent#removeallrows",
-    "class/src/components/editgrid/EditGrid.js~EditGridComponent.html#instance-method-removeAllRows",
-    "src/components/editgrid/EditGrid.js~EditGridComponent#removeAllRows",
-    "method"
   ],
   [
     "src/components/editgrid/editgrid.js~editgridcomponent#removerow",
@@ -9519,6 +9549,18 @@ window.esdocSearchIndex = [
     "src/components/form/form.js~formcomponent#_currentform",
     "class/src/components/form/Form.js~FormComponent.html#instance-member-_currentForm",
     "src/components/form/Form.js~FormComponent#_currentForm",
+    "member"
+  ],
+  [
+    "src/components/form/form.js~formcomponent#_parentvisible",
+    "class/src/components/form/Form.js~FormComponent.html#instance-member-_parentVisible",
+    "src/components/form/Form.js~FormComponent#_parentVisible",
+    "member"
+  ],
+  [
+    "src/components/form/form.js~formcomponent#_visible",
+    "class/src/components/form/Form.js~FormComponent.html#instance-member-_visible",
+    "src/components/form/Form.js~FormComponent#_visible",
     "member"
   ],
   [
@@ -12369,6 +12411,12 @@ window.esdocSearchIndex = [
     "src/components/textarea/editform/textarea.edit.display.js",
     "file/src/components/textarea/editForm/TextArea.edit.display.js.html",
     "src/components/textarea/editForm/TextArea.edit.display.js",
+    "file"
+  ],
+  [
+    "src/components/textarea/editform/textarea.edit.layout.js",
+    "file/src/components/textarea/editForm/TextArea.edit.layout.js.html",
+    "src/components/textarea/editForm/TextArea.edit.layout.js",
     "file"
   ],
   [