+++ src/components/button/Button.js
@@ -169,7 +169,7 @@ export default class ButtonComponent extends Field {
     this.on('change', (value, flags) => {
       const isValid = !(flags && flags.noValidate)
         ? (this.root
-            ? this.root.checkValidity(this.root.data)
+            ? this.root.checkValidity(this.root.data, null, null, true)
             : true)
         : value.isValid;
       this.loading = false;